+++ libc/src/sys/stat/linux/mkdir.cpp
@@ -20,9 +20,10 @@ namespace __llvm_libc {
 
 LLVM_LIBC_FUNCTION(int, mkdir, (const char *path, mode_t mode)) {
 #ifdef SYS_mkdir
-  long ret = __llvm_libc::syscall_impl(SYS_mkdir, path, mode);
+  int ret = static_cast<int>(__llvm_libc::syscall_impl(SYS_mkdir, path, mode));
 #elif defined(SYS_mkdirat)
-  long ret = __llvm_libc::syscall_impl(SYS_mkdirat, AT_FDCWD, path, mode);
+  int ret = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_mkdirat, AT_FDCWD, path, mode));
 #else
 #error "mkdir and mkdirat syscalls not available."
 #endif
