+++ libc/src/unistd/linux/pread.cpp
@@ -18,9 +18,10 @@ namespace __llvm_libc {
 
 LLVM_LIBC_FUNCTION(ssize_t, pread,
                    (int fd, void *buf, size_t count, off_t offset)) {
-  long ret = __llvm_libc::syscall_impl(SYS_pread64, fd, buf, count, offset);
+  ssize_t ret = static_cast<ssize_t>(
+      __llvm_libc::syscall_impl(SYS_pread64, fd, buf, count, offset));
   if (ret < 0) {
-    libc_errno = -ret;
+    libc_errno = static_cast<int>(-ret);
     return -1;
   }
   return ret;
